+++ src/main/java/at/hannibal2/skyhanni/features/inventory/chocolatefactory/ChocolateFactoryBarnManager.kt
@@ -9,6 +9,7 @@ import at.hannibal2.skyhanni.utils.NumberUtil.formatLong
 import at.hannibal2.skyhanni.utils.SimpleTimeMark
 import at.hannibal2.skyhanni.utils.SoundUtils
 import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
+import at.hannibal2.skyhanni.utils.TimeUtils.format
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import kotlin.time.Duration.Companion.seconds
 
@@ -50,7 +51,13 @@ object ChocolateFactoryBarnManager {
 
         rabbitDuplicatePattern.matchMatcher(event.message) {
             HoppityEggsManager.shareWaypointPrompt()
-            ChocolateAmount.addToAll(group("amount").formatLong())
+            val amount = group("amount").formatLong()
+            if (config.showDuplicateTime) {
+                val format = ChocolateFactoryAPI.timeUntilNeed(amount).format(maxUnits = 2)
+                event.chatComponent.appendText("\n§7(§a+§b$format §aof production§7)")
+                ChocolateAmount.averageChocPerSecond()
+            }
+            ChocolateAmount.addToAll(amount)
         }
 
         rabbitCrashedPattern.matchMatcher(event.message) {
